requirements.txt
PyOpenGL==3.1.0
PyQt5==5.11.3
numpy==1.15.3